[CHORE] Move _modelForMixin to Model package
