+++ packages/model/addon/-private/index.ts
@@ -3,3 +3,5 @@ export { default as belongsTo } from './belongs-to';
 export { default as hasMany } from './has-many';
 export { default as Model } from './model';
 export { default as Errors } from './errors';
+
+export { default as _modelForMixin } from './system/model-for-mixin';
